Add module updated file
